bit more reorganize...
