+++ src/platform/index.tsx
@@ -1,7 +1,7 @@
 /* @refresh reload */
 import { render } from 'solid-js/web'
 import 'lume'
-import Main from './app/Main.tsx'
+import Main from '@/app/Main.tsx'
 import '@/shared/styles/base.css'
 
 const root = document.getElementById('root')
