+++ src/platform/vite-env.d.ts
